+++ src/main/java/com/ticketmate/backend/test/controller/TestControllerDocs.java
@@ -54,7 +54,7 @@ public interface TestControllerDocs {
           - `INVALID_MEMBER_ROLE_REQUEST (400)`: 테스트 전용 Role(ROLE_TEST, ROLE_TEST_ADMIN) 외의 값이 요청될 경우
           """
   )
-  ResponseEntity<CompletableFuture<Void>> generateMockMembers(int count);
+  CompletableFuture<ResponseEntity<String>> generateMockMembers(int count);
 
   @Operation(
       summary = "테스트 회원 삭제",
@@ -87,7 +87,7 @@ public interface TestControllerDocs {
           
           """
   )
-  ResponseEntity<CompletableFuture<Void>> createConcertHallMockData(Integer count);
+  CompletableFuture<ResponseEntity<String>> createConcertHallMockData(Integer count);
 
   @Operation(
       summary = "테스트 공연 데이터 추가",
@@ -104,7 +104,7 @@ public interface TestControllerDocs {
           
           """
   )
-  ResponseEntity<CompletableFuture<Void>> createConcertMockData(Integer count);
+  CompletableFuture<ResponseEntity<String>> createConcertMockData(Integer count);
 
   @Operation(
       summary = "테스트 신청서 데이터 추가",
@@ -120,7 +120,7 @@ public interface TestControllerDocs {
           - 중복 데이터 생성에 따라 사용자가 원하는 개수보다 적은 데이터가 저장될 수 있습니다
           """
   )
-  ResponseEntity<CompletableFuture<Void>> createApplicationFormMockData(Integer count);
+  CompletableFuture<ResponseEntity<String>> createApplicationFormMockData(Integer count);
 
   @Operation(
       summary = "포트폴리오 Mock 데이터 비동기 생성",
@@ -142,5 +142,5 @@ public interface TestControllerDocs {
           - `INTERNAL_SERVER_ERROR (500)`: 포트폴리오 데이터 생성 또는 저장 중 서버 내부 오류가 발생한 경우
           """
   )
-  ResponseEntity<CompletableFuture<Void>> createPortfolioMockData(Integer count);
+  CompletableFuture<ResponseEntity<String>> createPortfolioMockData(Integer count);
 }
